Move Chat handling to UI project
