+++ src/openrct2-ui/interface/Chat.h
@@ -1,0 +1,51 @@
+/*****************************************************************************
+ * Copyright (c) 2014-2026 OpenRCT2 developers
+ *
+ * For a complete list of all authors, please refer to contributors.md
+ * Interested in contributing? Visit https://github.com/OpenRCT2/OpenRCT2
+ *
+ * OpenRCT2 is licensed under the GNU General Public License version 3.
+ *****************************************************************************/
+
+#pragma once
+
+#include <openrct2/core/StringTypes.h>
+#include <string_view>
+
+namespace OpenRCT2
+{
+    struct ColourWithFlags;
+
+    constexpr int8_t kChatHistorySize = 10;
+    constexpr int16_t kChatInputSize = 1024;
+    constexpr uint8_t kChatMaxMessageLength = 200;
+    constexpr int16_t kChatMaxWindowWidth = 600;
+
+    enum class ChatInput : uint8_t
+    {
+        none,
+        send,
+        close,
+    };
+
+    extern bool gChatOpen;
+
+    namespace Drawing
+    {
+        struct RenderTarget;
+    }
+
+    bool ChatAvailable();
+    void ChatOpen();
+    void ChatClose();
+    void ChatToggle();
+
+    void ChatInit();
+    void ChatUpdate();
+    void ChatDraw(Drawing::RenderTarget& rt, ColourWithFlags chatBackgroundColour);
+
+    void ChatAddHistory(std::string_view s);
+    void ChatInput(ChatInput input);
+
+    int32_t ChatStringWrappedGetHeight(u8string_view args, int32_t width);
+} // namespace OpenRCT2
