+++ src/openrct2-ui/libopenrct2ui.vcxproj
@@ -68,6 +68,7 @@
     <ClInclude Include="input\MouseInput.h" />
     <ClInclude Include="input\ShortcutIds.h" />
     <ClInclude Include="input\ShortcutManager.h" />
+    <ClInclude Include="interface\Chat.h" />
     <ClInclude Include="interface\Dropdown.h" />
     <ClInclude Include="interface\Graph.h" />
     <ClInclude Include="interface\FileBrowser.h" />
@@ -136,6 +137,7 @@
     <ClCompile Include="input\MouseInput.cpp" />
     <ClCompile Include="input\ShortcutInput.cpp" />
     <ClCompile Include="input\ShortcutManager.cpp" />
+    <ClCompile Include="interface\Chat.cpp" />
     <ClCompile Include="interface\Graph.cpp" />
     <ClCompile Include="interface\FileBrowser.cpp" />
     <ClCompile Include="interface\InGameConsole.cpp" />
