+++ src/openrct2-ui/windows/ServerStart.cpp
@@ -11,6 +11,7 @@
 
     #include "../interface/Theme.h"
 
+    #include <openrct2-ui/interface/Chat.h>
     #include <openrct2-ui/interface/Widget.h>
     #include <openrct2-ui/interface/Window.h>
     #include <openrct2-ui/windows/Windows.h>
@@ -19,7 +20,6 @@
     #include <openrct2/config/Config.h>
     #include <openrct2/core/String.hpp>
     #include <openrct2/drawing/Text.h>
-    #include <openrct2/interface/Chat.h>
     #include <openrct2/network/Network.h>
     #include <openrct2/ui/WindowManager.h>
     #include <openrct2/windows/Intent.h>
