+++ src/openrct2/libopenrct2.vcxproj
@@ -299,7 +299,6 @@
     <ClInclude Include="GameStateSnapshots.h" />
     <ClInclude Include="Identifiers.h" />
     <ClInclude Include="Input.h" />
-    <ClInclude Include="interface\Chat.h" />
     <ClInclude Include="interface\ColourWithFlags.h" />
     <ClInclude Include="interface\Cursors.h" />
     <ClInclude Include="interface\FontFamilies.h" />
@@ -943,7 +942,6 @@
     <ClCompile Include="GameState.cpp" />
     <ClCompile Include="GameStateSnapshots.cpp" />
     <ClCompile Include="Input.cpp" />
-    <ClCompile Include="interface\Chat.cpp" />
     <ClCompile Include="interface\ColourWithFlags.cpp" />
     <ClCompile Include="interface\Cursors.cpp" />
     <ClCompile Include="interface\FontFamilies.cpp" />
