+++ src/openrct2/windows/Intent.h
@@ -58,6 +58,9 @@ namespace OpenRCT2
         INTENT_ACTION_REMOVE_PROVISIONAL_FOOTPATH,
         INTENT_ACTION_REMOVE_PROVISIONAL_TRACK_PIECE,
         INTENT_ACTION_REFRESH_PLAYER_LIST,
+        INTENT_ACTION_INIT_CHAT,
+        INTENT_ACTION_UPDATE_CHAT,
+        INTENT_ACTION_CHAT_ADD_HISTORY,
 
         INTENT_ACTION_NULL = 255,
     };
